backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
openai
chromadb
sentence-transformers
assemblyai

